// modules/loan/modules/third-party-loans/enum/own-third-party-loans-control-name.enum.ts
export enum AttributeOwnThirdPartyLoansTable {
  FILTER = 'filter'
}

export enum OwnThirdPartyLoansTableKeys {

  ALIAS = 'alias',
  IDENTIFIER = 'identifier',
  CURRENCY = 'currency',
  NAME = 'name',
  TOTAL_DEBT = 'totalAmount',
  TYPE_ACCOUNT = 'productLabel',
  TYPE = 'type'


}

export enum OwnThirdPartyLoansTypeProduct {
  LOANS = '04',
  PRODUCT = 4
}

export enum AttributeOwnLoansTable {
  FILTER = 'filter'
}